feat: add shadow dom support for base styles (#75)
## Description

The point of this is to account for luminous being used in shadow roots.

By using `getRootNode` (if supported), we can get to the `#shadow-root` rather than `document`. If we're not in one, we get `document`.

### New Feature

- [ ] If this is a big feature with breaking changes, consider [opening an issue][issues] to discuss first. This is completely up to you, but please keep in mind that your PR might not be accepted.
- [x] Run unit tests to ensure all existing tests are still passing
- [x] Add new passing unit tests to cover the code introduced by your PR
- [ ] Update the readme
- [x] Add some [steps](#steps-to-test) so we can test your cool new feature!
- [x] The PR title is in the [conventional commit](#conventional-commit-spec) format: e.g. `feat(<area>): added new way to do this cool thing #issue-number`
- [x] Add your info to the [contributors](#packagejson-contributors) array in package.json!

## Steps to Test

Create an element just like in the added test. Its as simple as your target element for luminous being a child of a shadow root.