build: enable semantic release on all branches
This commit changes the travis release script to release on all
branches. It also changes the replace plugin to a regex instead of a
string, which is more reliable and matches the release plugin's docs
more closely. It also fixes an issue where the replacement files
directory was missing the `js` subdirectory.